headless tests: Move `run-headless.js` to its own file and `include_str!` it
